switch to ChunkCodecLibLz4
