+++ Project.toml
@@ -13,24 +13,24 @@ ScopedValues = "7e506255-f358-4e82-b7e4-beb19740aa63"
 
 [weakdeps]
 Blosc = "a74b3585-a348-5f62-a45c-50e91977d574"
+ChunkCodecLibLz4 = "7e9cc85e-5614-42a3-ad86-b78f920b38a5"
 CodecBzip2 = "523fee87-0ab8-5b00-afb7-3ecf72e48cfd"
-CodecLz4 = "5ba52731-8f18-5e0d-9241-30f10d1ec561"
 CodecZstd = "6b39b394-51ab-5f42-8807-6242bab2b4c2"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 bitshuffle_jll = "228fe19c-1b83-5282-a626-13744502a320"
 
 [extensions]
 BloscExt = "Blosc"
+Lz4Ext = "ChunkCodecLibLz4"
 CodecBzip2Ext = "CodecBzip2"
-CodecLz4Ext = "CodecLz4"
 CodecZstdExt = "CodecZstd"
 UnPackExt = "UnPack"
 bitshuffle_jllExt = "bitshuffle_jll"
 
 [compat]
 Blosc = "0.7.3"
+ChunkCodecLibLz4 = "0.2.1"
 CodecBzip2 = "0.8.5"
-CodecLz4 = "0.4.6"
 CodecZlib = "0.7.8"
 CodecZstd = "0.8"
 FileIO = "1.5"
