--- ext/CodecLz4Ext/THIRDPARTY.md
@@ -1,71 +1,0 @@
-JLD2/CodecLz4Ext.jl is derived from HDF5.jl/CodecLz4Ext.jl with the MIT license.
-It itself was derired from H5Zlz4.c from the HDF5 Group with license also below.
-
-
-The MIT License (MIT)
-Copyright (c) 2012-2021: Timothy E. Holy, Simon Kornblith, and contributors: https://github.com/JuliaIO/HDF5.jl/contributors
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software
-and associated documentation files (the "Software"), to deal in the Software without
-restriction, including without limitation the rights to use, copy, modify, merge, publish,
-distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or
-substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
-BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-
----
-
-
-HDF5.jl/CodecLz4Ext.jl is derived from H5Zlz4.c from HDF5 Group
-
-================================================================================
-LZ4 filter plugin license
-================================================================================
-
-Copyright Notice and License Terms for
-HDF5 LZ4 compression filter plugin
------------------------------------------------------------------------------
-
-HDF5 LZ4 compression filter plugin
-Copyright 2013-2015 by The HDF Group.
-
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted for any purpose (including commercial purposes)
-provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice,
-   this list of conditions, and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright notice,
-   this list of conditions, and the following disclaimer in the documentation
-   and/or materials provided with the distribution.
-
-3. In addition, redistributions of modified forms of the source or binary
-   code must carry prominent notices stating that the original code was
-   changed and the date of the change.
-
-4. All publications or advertising materials mentioning features or use of
-   this software are asked, but not required, to acknowledge that it was
-   developed by The HDF Group and credit the contributors.
-
-5. Neither the name of The HDF Group, nor the name of any Contributor may
-   be used to endorse or promote products derived from this software
-   without specific prior written permission from The HDF Group or the
-   Contributor, respectively.
-
-DISCLAIMER:
-THIS SOFTWARE IS PROVIDED BY THE HDF GROUP AND THE CONTRIBUTORS
-"AS IS" WITH NO WARRANTY OF ANY KIND, EITHER EXPRESSED OR IMPLIED.  In no
-event shall The HDF Group or the Contributors be liable for any damages
-suffered by the users arising out of the use of this software, even if
-advised of the possibility of such damage.
