+++ filterpkgs/JLD2Lz4/Project.toml
@@ -4,9 +4,8 @@ authors = ["Jonas Isensee <jonas.isensee@web.de>"]
 version = "0.1.0"
 
 [deps]
-CodecLz4 = "5ba52731-8f18-5e0d-9241-30f10d1ec561"
+ChunkCodecLibLz4 = "7e9cc85e-5614-42a3-ad86-b78f920b38a5"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 
 [compat]
-CodecLz4 = "0.4.6"
 JLD2 = "0.5.15"
